+++ src/main/java/com/emarques/seniorapi/api/model/ouput/PedidoOutput.java
@@ -22,7 +22,7 @@ public class PedidoOutput {
 
     private BigDecimal valorTotalComDesconto;
 
-    private Integer desconto;
+    private BigDecimal desconto;
 
     private Endereco enderecoEntrega;
